// enums.ts
export enum Region {
  England = 'ENGLAND',
  NorthernIreland = 'N.IRELAND',
  Scotland = 'SCOTLAND',
  Wales = 'WALES'
}

export enum PropertyStatus {
  PropertyVacant = 'PROPERTY_VACANT',
  PartiallyVacant = 'PARTIALLY_VACANT',
  PropertyActive = 'PROPERTY_ACTIVE',
  PropertyOverdue = 'PROPERTY_OVERDUE'
}